Run cfn-hup via a cron job.
cfn-hup is supposed to be a daemon, but it doesn't work that way. So run
it via a wrapper in a cron job. Also run this way, we won't have
logging, so make sure os-refresh-config is run via upstart so we get the
output logged.

The wrapper must use a lock to avoid problems that might be caused when
multiple updates overlap.

Change-Id: I8613c45b33091ad6448ecddebe6a1f4337f09d7d