Avoid race condition with 1 second sleep.
Upstart will return from 'start xxx' immediately upon forking. We still
need to run start-stop-daemon, change users, exec again, and then run
the python code to get to listen() before the service is actually ready.
This all should take < 1s, but the sleep is just a stopgap until we have
a better way to track "is the service ready" in upstart.

Fixes bug #1179766

Change-Id: I3842e572c940e833cb3036b67df0ee71fe5bb129