Work around boto dependency in heat-cfntools.
Heat-cfntools has a hard dependency on boto == 2.5.2. However, this
version of boto breaks newer versions of cloud-init when used with
OpenStack and specifically with Heat generated userdata. That problem is
documented here: http://pad.lv/1177576

The change to allow >= 2.5.2 has been submitted to be fixed upstream
and can be tracked in this bug: http://pad.lv/1180582 .

Change-Id: I1ce96901754472e6b0d518d26431d3b3a26daea6