Add post-start option to os-svc-daemon.
We also migrate os-svc-daemon to supporting switches to modify its
behavior and be more flexible as it grows. The old positional style is
still supported but will be deprecated.

Each service will have a unique way to verify that it has started. The
default 'exec sleep 1' will be sufficient for some, but not all. Quantum
is the first example, which can take more than 5 seconds to open a
listening socket.

Due to a lack of knowledge and testing on systemd, it will simply print
out a warning about the issue until we have a better answer.

Change-Id: I8d0e0842dfa66b2e32b942270ed071e593d3865e