Add respawn limit to prevent infinite respawn.
The respawn count is not incremented until post-start exits, so we
need to set the count lower than the interval to allow for the lag
caused by the one second sleep. Two times in 5 seconds seems like
enough to handle the occasional unexpected crash but stay dead when
things are badly broken.

Change-Id: I8eaca6f9ea131861a42d473ffc5c09a41226ce0e