Make quantum-server post-start sleep 10 seconds.
A better long term solution is something that polls trying to reach
quantum-server on the port it is configured to listen on. However,
extending the usual sleep to 10 seconds has at least a moderate chance
of avoiding the race between 'start quantum-server' and the listen()
call that makes it reachable.

Change-Id: I56453e3e9d718216e3cfee0a3bc353c51f74b3be