Set pipefail on init-quantum to detect failures.
Even with the sleeps in the upstart jobs, sometimes we still lose the
race with quantum-server. This makes quantum net-create fail. We should
stop init-quantum right then, but because it is being piped to other
commands, the failure is not detected.

Pipefail will make sure that we detect any of these failures and stop
the process.

Change-Id: I9f5f3863903aa6abd816841dd71133d24ffe6358