Automate network router setup.
If quantum.ovs.fixed_range is defined, that network will need to have a
router attached in order to reach the ec2 metadata service, or to reach
any floating ip ranges.

If quantum.floatingip_range is defined, that subnet will be created and
assigned to the default router. This is independent of openvswitch,
which is why it is not in the quantum.ovs namespace.

Change-Id: I7831be7435505885c77c9eeef33306eee546ac98