Make 01-boot-stack support requirements.txt.
Updates install.d/01-boot-stack so that it looks in requirements.txt
as well as the older (legacy) tools/pip-requires files when
attempting to install dependencies for our client libraries.

This fixes a rather cryptic issue that crept in this morning
with quantumclient due to the fact that they just moved over
to use requirements.txt... this would essentially cause some
of the quantumclient dependencies to get incorrectly easy_install'ed and
thus have bad permissions (on some of the egg-info files).

Change-Id: I488be161c45408727adeceb26f1f5c6b1de01430