Remove boot-stacks dependency on openstack-db
Everything provided by the openstack-db element is already done in the
boot-stack elemnt or one of it dependencies. So it shouldn't be included
as a dependency.

The only piece that is missing is the line the installs the cinder
service because boot-stack doesn't depend on cinder. The cinder element
currently seems to not be working. So I'm placing that line to install the
cinder service in the boot-stack element, this can be removed again once
the cinder element is working.

This will allow me to make progress on using the new source-interface,
as multiple elements will no longer be adding the same service.

Change-Id: I808a580cb8731c83fff6dd50e72db4a29150a74e