Add "install from directory" to os-svc-install
This will allow elements to install from a directory that has
been downloaded by the source-repositories element.

In cases where this install method is used no attempt is made
to verify the directory being installed from. The assumption is
that the source-repositories element errors if multiple elements
try to download to the same directory. Its also assumed that the
repository version downloaded is correct as this can now be set
by the process running d-i-b and the element shouldn't be aware
of it.

Change-Id: Iceada839f0418789156741b56979c4a287a9a24d