Remove boot-stack dependency on cinder
Cinder is not used in all boot-stack environments, only in the
overcloud. We may want to have separate elements to initialise cinder
in the future, but that is speculative. Right now we only know that we
don't need cinder except when building overcloud images.

This was prompted by bug 1199412 occuring.

Change-Id: Ie7485f741aaef67599fb1fa2da381e23b6cbe805