Restart keystone in wipe-openstack
Fixes bug 1199568

Depending on the timing on first boot if keystone is started
before os-refresh-config then the file /etc/keystone/logging.conf
will not be in place yet. Causing keystone to fail to start, needing
to be started again after os-config-applier is run but before
init-keystone.

After some recent changes, this is now consistently happening on Fedora.

Change-Id: Ifc979bad0ac351794a12ccda2dd97f277b79721b