Quantum -> Neutron rename
Rename all paths to neutron, and almost all file entries,
except some of those in the nova.conf template(these have not been
ported to neutron yet).

This does not "Fix" bug 1199443 more accurately it avoids it. But
a rename would eventually be needed in tripleo anyway so doing
this now means we don't need to find/wait for a fix.

Fixes bug 1199443

Change-Id: I5e61fab1d444d90eb2c54ccfeefca5e148d96996