Remove os-refresh-config call from wipe-openstack.
Previously we were running os-refresh-config to make sure we
caught everything that had been refactored into orc scripts. After
analysis, it seems that the only thing we need before init-neutron is
init-neutron-ovs. This is called later in the orc chain, so we have
split wipe-openstack's call to init-neutron out into 85-init-neutron.

In addition, there was a check for "is it already running" behavior that
is no longer needed since we should only have one orc running at a time
with the impending lock change. We have removed that behavior.

Finally the init/wipe scripts don't need to be in the system PATH and thus
have been merged with their orc scripts, which were their only callers.

Change-Id: I407678104353b180102e16894408c3ffa2d7db90