Improve snmpd element.
The purpose of this patch is to improve the ability to monitor hosts,
correct an error in the current snmpd.conf, and align variable names
with their function.

Renames the Heat variable name prefix from "snmp" to "snmpd" in order to
align the Heat variable name with the element name.

Changes the Heat variable MIB to export_MIB to better represent what
the variable is used for.

Removes a hardcoded user name in snmpd.conf for the rouser to fix
an authentication issue for user names other then "authOnlyUser".

The OoOEnabled command will allow elements like Nagios to poll
hosts with this element via SNMP, and receive a consistent
response.

The CheckProcOpenFiles command is a simple Nagios style check of
open files per process that will be monitored so that an alert can
be sent if any one process exceeds its open file limit.

Change-Id: I92afd1f81dfc47c3bb0f4bb3a5aa72bee8870f48
Authored-by: Chris Krelle <nobodycam@gmail.com>