Install os-apply-config in a virtualenv.
Due to the older python distribute still shipping on Fedora, Ubuntu,
and other distros, os-apply-config can not be installed with pip
install -U.  This change installs os-apply-config in a virtualenv
instead, so that it's in an isolated environment.  os-apply-config and
os-config-applier are symlinked under /usr/local/bin so that all calls
continue to work.

Change-Id: I50d7147cf74a36bbe7751d545c353ac354bdd189