Update network configuration for overclouds.
The overcloud network configuration was not migrating ip addresses and
ports onto the exterior integration bridge. This is fine when the
bridge is not the default route, and/or is on a different interface,
but behaves very poorly when only a single interface is present. Since
we don't use br-ex in undercloud setups (today), and don't use
br-ctlplane in overcloud setups, we can use the same code for
migrating address and routes by updating our configuration instead.

Change-Id: I85a455bd2b03089bf39c34a3b69af78af09ce286