Align build files and requirements
This is not an actual python project, so there is an amount of cruft
in the repo that is not needed. However, it does generate tarballs in
OpenStack Infra, so it needs the tox venv environment and the pbr
setup.py stuff. It also is totally possible to check python files in
the repo for flake8 violations, so set that up.

Change-Id: Ifb94f65912716efe1c4596d552beee92d2dc3d6a