Stop doing complex keystone intialisation.
Initialisation tasks which are one-time should be API orchestration
rather than instance level orchestration : doing otherwise makes
instance bring up a mass of special cases, and reduces flexability
below that that the base API offers. We're migration all things to
this pattern, keystone is merely the last in a line of fixes :).

Change-Id: I125a1c742289b3d1cf6319d127f4cf948459d9bc