Remove debug and verbose config from nova.conf
Verbose is overridden by the log_config parameter, since
this is overridden anyway no reason to keep it around as it can
only lead to confusion.

Debug is used to set the log level, although its overridden by
log_config as well, and is also used to turn on debug mode only checks.
Since we want the default values here to be for production we shouldn't
set debug=True here.

Change-Id: I003ea1a80efc3e9271015393feb7ef7413412331