Add an element installing tuskar API service
This should make it easier to deploy tuskar in the undercloud
(basically, one will just add 'tuskar' the the list of elements
when building a disk image for the undercloud node).

This depends on Ib66e688bd7500f03dea7d4282607cd51078e9a51 to be
merged into tripleo-heat-templates to work properly.

Change-Id: I1d16e535ae50a06a8792e1613bb6b6c4647bf128