Update OpenStack services to support qpidd element
Configuration files are changed so that qpid is used as the rpc_backend
instead of kombu (rabbitmq) if it is specified in the metadata.

Add qpid-python to Heat, Neutron, and Cinder requirements.txt file.
These elements install using venv and don't use the system installed
libraries. These changes can be removed if requirements.txt is changed
upstream. There are two bugs tracking these changes:
https://bugs.launchpad.net/heat/+bug/1225191
https://bugs.launchpad.net/neutron/+bug/1225232

Change-Id: I9bf6ab9d3bae668e592af65b5c8075a487e07261