Support configuring a VLAN on the public network.
The CD overcloud needs a VLAN on the neutron network node, or we can't
get public traffic to VM's in or out. This isn't an entirely general
solution but that is something we should be looking to Neutron and DHCP for,
not manual metadata: as a stop gap this should be sufficient.

Change-Id: Ifaed03b96de12453bcf23cdcf2e91bb4dc32349c