Setup overcloud user and administrative accounts.
To avoid polluting incubator with the cd cloud specific knowledge,
like the networking, we pass in the list of users and administrators.
Note that we could just make this hidden, but that makes it harder to
introspect. The incubator code will consider these lists authoritative
and just consult deploy-environment state for the passwords to inject
for users without accounts in the cloud.

Change-Id: I1a6cf06d1a72f4fee6ab198ab4db5d803fbb7202