Adding ml2 support to baremetal.
Changes the neutron plugin and router to neutron ml2. The Modular
Layer 2 (ml2) plugin is a framework allowing OpenStack Networking
to simultaneously utilize the variety of layer 2 networking
technologies. It currently works with the existing openvswitch,
linuxbridge, and hyperv L2 agents, and is intended to replace
and deprecate the monolithic plugins associated with those L2
agents.

Change-Id: I4aff2f647c8e0b0ae4014467290482e9ca351386