Switch novato the upstream api-paste.ini
We had a copy of this file which needed to be maintained by tripleo,
using the upstream version means we should no longer have to maintain
it. In this case the reason for doing this is to enable the nova v3 API.

The tempest tests for the nova v3 api require it to be enabled.

Change-Id: I98583009cd18adda4443ce0b00a584434a4abf81