Update nova-kvm to destroy the default libvirt net
Remove files to start/create the default libvirt network.

This is important because the default libvirt network in the overcloud
may clash with the libvirt network used to host the seed VM (unless
the user has modified their local libvirt network to use a different
range).

This can cause issues when trying to ping or access overcloud-compute
(or overcloud-all-in-one) from outside of the seed VM.

For most people this should make overcloud-compute accessible from the
machine hosting the seed.

Change-Id: I40ae9e61367481fcc7bb46cb27a7334527b77f0c