Add element for tripleo-heat-templates.
Installs tripleo-heat-templates from git under /opt/stack using the
source-repositories element.  The tuskar element is currently doing this
on it's own in order to get access to the templates, and I'd like to see
it in a seperate element instead.  Other consumers may want access to
the templates outside of using tuskar.  Will submit another change to
update the tuskar element to use this element once merged.

Change-Id: Ibfc61f4f677f051f409cbfef571c1da67356e440