Fix use-ephemeral rsync of directory
If the stateful path source is a directory we need to copy the directory
completely to the target, otherwise source directory will be copied as a
sub-directory of the target. For example /var/lib/cinder will be copied
to /mnt/state/var/lib/cinder/cinder.

We also fix bugs with doubling /'s which may cause problems with the
script later on.

Change-Id: I51736171db6059e30f147c5027866a538305f0f5