Nova: make logging configurable...
Adds top level nova.debug and nova.verbose config parameters
to control Nova logging.

The motivation for this change is so that we can more easily
control the log level of Nova. For production our default log
level is still WARN, and for CI we'll be able to easily increase
the level to DEBUG.

Change-Id: Ia24a6021dd42eae9719a642f639346f217c92f30