Keystone: make logging configurable...
Adds top level keystone.debug and keystone.verbose config parameters
to control Keystone logging.

The motivation for this change is so that we can more easily
control the log level of Keystone. For production our default log
level is still WARN, and for CI we'll be able to easily increase
the level to DEBUG.

Change-Id: Id164b484c425f1a20c12aebbea5316702190e2f4