Heat: make logging configurable...
Adds top level heat.debug and heat.verbose config parameters
to control Heat logging.

The motivation for this change is so that we can more easily
control the log level of Heat. For production our default log
level is still WARN, and for CI we'll be able to easily increase
the level to DEBUG.

Change-Id: Ief8fc8c11532ae3f122fed97047b231af0f1e914