Cinder: make logging configurable...
Adds top level cinder.debug and cinder.verbose config parameters
to control Cinder logging.

The motivation for this change is so that we can more easily
control the log level of Cinder. For production our default log
level is still WARN, and for CI we'll be able to easily increase
the level to DEBUG.

Change-Id: I637a670182c9c5f4e15cd6699857216c2aa13b36