Neutron: make logging configurable...
Adds top level neutron.debug and neutron.verbose config parameters
to control Neutron logging.

The motivation for this change is so that we can more easily
control the log level of Neutron. For production our default log
level is still WARN, and for CI we'll be able to easily increase
the level to DEBUG.

Change-Id: I6e02be47db485037fe2a28cb30f0c7253affc20c