Use [client] instead of [mysql] in my.cnf
The [mysql] group technically only informs the command line client. We
should be using the [client] group which will inform libmysqlclient in
general. Note that there are still going to be programs that bypass this
file so it is not a perfect solution to the problem of different
binaries having different compiled in defaults for the socket.

Change-Id: Ic82a976b8591ef3a19e45f36ecc210104e670734