Updates cinder element
The previous cinder element was lacking a notty file and didn't
open the required ports in iptables.  This patch also allows a
developer to install the minimum required components for a
single cinder instance or (n) cinder-volume nodes.

This update results in 3 elements
cinder        : common cinder element
cinder-api    : installs the cinder api and scheduler
cinder-volume : installs the cinder volume service

Change-Id: I1ed9f67848be5716135085b6eaebc1f381ad83f9