Add new neutron-ovs-cleanup element.
This patch adds a new neutron-ovs-cleanup element which
configures the neutron-ovs-cleanup service to run
so that it cleans up *all* OVS ports. We want to do this
at a very specific time in the boot process so that DHCP
on reboot works properly:

 1) after OVS is available (nonetworking)
 2) before networking is started

NOTE: This service is not started on first boot but will
run on any subsequent reboots (it is enabled in os-refresh-config)

Without this fix networking after rebooting an overcloud instance
is a total failure due to the fact that the physical NIC has been
moved to an OVS bridge by init-neutron-ovs and DHCP requests
fail.

The currently patch implements this on systemd only.

In addition to all the above, cleaning up old neutron ports is
a good idea on reboot anyways. So we get that for free here too.

Change-Id: Ic87f2388680ec4f67b6275b8d923f76b75559c78
Closes-bug: 1254533