Add nova-ironic element to replace nova-baremetal
Add new nova-ironic element which can be used in place of the existing
nova-baremetal element. These are not meant to be used together. This
patch also moves the Ironic ssh power key to /mnt/state.

NOTE: Nova "ironic" driver has not landed upstream yet, but this image
can be built with it ahead of time by setting the following:

  export DIB_REPOLOCATION_nova=https://review.openstack.org/openstack/nova
  export DIB_REPOREF_nova=refs/changes/28/51328

Change-Id: Ia8a61bf55f4337f40ee40b0fd381b2f776f67312