Remove --system-site-packages from clients virtualenv
Using --system-site-packages here is causing us to use old versions of
libraries even though the correct version is installed into the
virtualenv, this is due to the order of the various directories in
sys.path. This is not a problem on newer versions of
virtualenv/setuptools although I couldn't find out exactly when it was
fixed.

In this case the specific conflict was with python-six. A recent
unrelated commit caused us to stop pip upgrading the version installed
from a package. In turn breaking several of the virtualenv installed
packages.

Change-Id: I030b57c1c3ece3772028ad08cc506c39bb8f3e21