Add os-svc-restart
Add os-svc-restart, a helper script to restart a service. It uses
map-services to figure out the actual service name, as the service name
can differ based on the distro being used and also if installing from
source or packages.  Currently, only if the init system is systemd will
a call out to map-services be done. This matches the pattern from
install-packages where map-packages is not used on dpkg systems since
those package names are used as the default in the source install
scripts.

Having a helper script will avoid the proliferation of the following
pattern everywhere:

service foo restart || service foo-by-another-name restart

This also significantly reduces errors in the log due to the above,
which can be misleading.

Depends on Change-Id: Ie7b2dcfa7cabd887d5c212df110d90f9d00a7f65

Change-Id: I15eb2a5db4b0a08e1fb40bda640cd8f224939a92