Add package install support to openstack-clients
Each client has its own install type directory. This is needed because
of the multiple repositories with different names in
source-repository-clients.

The client install logic has been refactored into the
install-openstack-client script so that each source install can reuse
it.

An alternative implementation would be add the ability to the
source-repositories element to support an alternate repo override name
for elements that have multiple entries in their source-repository
configuration. For example, the source-repository file for
openstack-clients is called source-repository-clients. We could add
support to look for a clients-source-install or clients-package-install
directory. However, we'd lose the ability to mix source and package
install types from the same element. Given that, I went with the
implementation here.

Note that the ironic client is not broken out by install type, because
there are no packages that yet exist for it.

Change-Id: I2700490d139db47b69dd0f30cd46d9d7e72b7809