Fix rabbitmq-server script on Ubuntu
This script was failing due to errors but not stopping because it was
not running with "set -e". This caused the old broken sysvinit script
from the Debian/Ubuntu packages to be used, which presumably regressed
an old bug around restarting, though it is not known.

There is no good reason to run with /bin/sh as the interpreter and we
are used to adding bashisms, so we convert the script to bash and add
set -e to protect against future sojourns into bash-land.

Change-Id: Ib334f336f3483ed40283e1931c9b023e572fffa2