Revert "Add package install support to openstack-clients"
There was a problem with this commit that causes a regression, I added
a +A when I intended to -1. This needs to be reverted in order to allow
time to fix. Essentially the clients arn't getting installed on the
images because the reponame varies for each client library and
source-repositories does not support this.

This reverts commit 74438c32e0f8c29600aa57656555c88e69106afd.

Change-Id: Id24426027b3043538e41b4bddbe45750d78dbfb6