Install packages required to build clients
These were dropped when the package support was added, they are not
required for package builds but for now there is no way to select an
element wide install script based on install type. So we install in
all cases.

Also adding 3 new dependencies(python-virtualenv, libxml2-dev and
libxslt-dev) that were never listed but are required.

Closes-Bug: #1274846
Change-Id: I8eb04b6d1527de43b768e2b51a1c916202ea5ee9