Fixes heat-cfntools install
When using heat-cfntools with diskimage-builder, the pip install
of heat-cfntools package is skipped because it's already in the
system site-packages directory.  Unfortunately it doesn't copy
the scripts from the heat-cfntools bin directory into the venv
bin directory and the diskimage-builder command will fail.

This patch adds gcc to the install-packages line and removes
the --system-site-packages flag from the virtualenv creation.
This ensures the heat-cfntools package is properly installed.

Change-Id: I29d8f310a47ea2b5d4148cf3fa32485dea3424e8