Fixes tgtd restart in cinder element
In a previous patch, the restart of tgtd was erroneously removed
and causes the cinder-volume service to create a symlink to a
volume but return an error to the cinder-scheduler.  This "volume"
must then be manually removed and tgtd service restarted before
cinder-volume works properly.

This patch adds the restart of tgtd again and allows cinder to
work as advertised without additional intervention.

Change-Id: I1c2cb22773ab8b39e4bbe028474665df81ac08a1