Install syslinux also in nova-baremetal
Previously the element relied on nova installing syslinux
before implicitly, which was not always the case in package
install mode. Also install it where it is actually used. Now
it will be installed twice, which should normally not hurt
(works in my testing). A previous iteration with a mere move
was rejected (https://review.openstack.org/#/c/70718/).

Change-Id: I8b460beda4275eff5878031593ce3c15307e94c9