Fix neutron package install
Previously the neutron element was installing neutron-server. This
should actually be happening in the neutron-server element itself.
Likewise, the neutron-dhcp-agent should also install the package it
needs. In both cases, the package is just openstack-neutron, but only
the intended services will be enabled and started due to the
os-refresh-config scripts. E.g., if you only include neutron-dhcp-agent,
and not neuton-server, only the neutron-dhcp-agent will be enabled.

Change-Id: I11e52e9cdd6062b03d389971acad1c52a9cc3ea8