Move cinder-volume creation to post-configure
This is actually not what migration is for. If something must happen
before a service is started, that is a post configuration thing that we
should order before enabling/starting/restarting the service.

Os-refresh-config will be changed to run migration after post-configure,
since it is for on-line things, not offline things.

Change-Id: I2c4eccab3010ec91fcd8129951f638b37bc0932a