Adds nfs-server element
This patch adds an element that installs and configures nfs
services for testing the cinder nfs backend in an environment
where no nfs services are present (e.g. CI)

When added to an image, this element creates a directory at
/mnt/state/var/lib/nfs and shares it to each node of the
overcloud.

This patch that handles the package mapping in diskimage-builder
needs to be applied first: https://review.openstack.org/#/c/76398/.

Change-Id: I4c425af64b866caa80b32b808e8a3b97d71ca81d