Install latest version of pip and virtualenv
We install pip with get-pip.py and use this to install virtualenv. This
is done as part of the new pip-and-virtualenv element. All elements that
need pip or virtualenv should use this element to install them.

Our motivation is that we need to talk to a pypi mirror generated with
a recent http://github.com/openstack-infra/pypi-mirror This mirror
caches some 'wheels' that the previous version of pip & virtualenv
can't install.

virtualenv contains its own version of pip that is used for all
virtualenv environments.

Change-Id: I282fd8fffc8d5707a078f78f2b4571138d7266f3