nova-kvm: configure /etc/sysconfig/libvirtd
Adds the --listen option to the /etc/sysconfig/libvirtd file.
This would be required for migrations on distributions that use
this file.

This commit also fixes an issue w/ a21a2e7 where the debian
style file was copied into place on *all* distributions.
We should avoid writing unnecessary files to disk where
possible. As part of this fix the file was moved from being
configured w/ os-config-apply into the install.d. It is unlikely
these files will change often so I think this is Okay...

Change-Id: I3725897744c13184d72475e137ff9267ee5449fc
Closes-bug: #1279356