Install python-keystoneclient in the swiftclient venv
The swift client can be used without keystone auth so keystoneclient
isn't listed as a dependency. We are using keystone so need to
explicitly include it.

A fix has also been submited to python-swiftclient to avoid the need for
this workaround I88f4934811b86b2995d722c60f9b6496f986e52a

Change-Id: Ic13d66fbe1e459cb3b17dab293f56e82806b9355