Link db sync utilities to /usr/local/bin...
Updates the ceilometer, ironic, neutron, and tuskar elements
so that we soft link the DB sync utilities into /usr/local/bin.

This allows us to refactor the reset-db script to avoid hard
coding the /opt/stack/venv binary patch which is used for
source installs.

This is cleaner and makes things consistent. It also makes
the source installs consistent with how most distro packages
work (thus making reset-db useful with packages as well).

Change-Id: Ib8c36811230512e4283f3f960254b2c37e699f46