Nova: don't hard code the venv dir
Updates the nova* elements so that we use a new
NOVA_VENV_DIR variable (set via an environment.d hook)
to configure the virtualenv install directory.

This is cleaner in that we aren't hard coding /opt/stack/venv
everywhere and also supports both isolated and shared
venvs (which some people would like to use).

Change-Id: I9b2ab8fd5e51b2530ec7a2b49bcc10aa369b0288