Neutron: don't hard code the venv dir
Updates the neutron* elements so that we use a new
NEUTRON_VENV_DIR variable (set via an environment.d hook)
to configure the installation directory for git installs.

This is cleaner in that we aren't hard coding /opt/stack/venv
everywhere and also supports both isolated and shared
venvs (which some people would like to use).

Change-Id: Ib9bf3ba32ff4da798f5e9a776bd75c7e321fd543