Add -i options to os-svc-install/os-svc-daemon...
Update os-svc-install with a new -i option which can be used
to specify the virtualenv directory.

Update os-svc-install with a new -s option which can be used
to enable system-site-packages within the virtualenv. By
default this is disabled (matches virtualenv's default...)
In doing this we remove a hard coded check for the 'nova' service
by making nova use the -s option instead.

Update os-svc-daemon with a new -i option which can be used
to specify the virtualenv directory.

NOTE:
 I used -i for both os-svc-daemon and os-svc-install since it wasn't
 already used and I wanted to make the option the same in both places.

Change-Id: Ief53567711b0153de890d50a28c471b1e5acd56c