Allow os-svc-daemon to specify dir name with -d option
In some circumstances we need to be able to create a /var/run
directory without actually installing a service.  This behavior
was supported before, but these changes allow the user to specify
the directory name and streamline the process by making -c an
optional parameter instead of requiring a dummy value.  It also
makes this functionality available on upstart-based systems.

The existing uses of -d had to be updated as well to reflect this
new behavior.

Change-Id: I34f33f2290eeefb5be82c0cf5198ae53d34332ce