Add mysql-mariadb-compat element
This allows distributions that have dropped the mysql aliases from
their mariadb packages to continue using the mysql element by
creating the necessary /var/run/mysqld directory.

At the moment this is only for RHEL7, but in the future it is
likely that Fedora will need the same thing because apparently this
is the approved direction for the mariadb package.

Change-Id: I2cc664e92b14db6bfdc2f33d0360001851e8ea1c