Factor out tgt-specific parts of cinder element
RHEL is only going to support the lio iscsi helper, so the tgt
parts of the element are not going to work there.  A subsequent
change will add a corresponding cinder-lio element to support that.

When this merges, the corresponding change to tripleo-incubator in
I1e33af0cb78fa9477bc5a0b9b89b06b907b81fae must merge as well to
prevent breakage of devtest.

Change-Id: Ic032204076adfa0d30b09423f2a46b587e07400d