Openstack-clients: don't hard code venv
Updates install-openstack-clients so that it parses args rather than
using positional parameters.

Updates the various source install elements so that we use a
new {CLIENTNAME}_VENV_DIR variable (set via an environment.d hook)
to configure the virtualenv directory. The in element default is
the same as before (/opt/stack/venvs/python-clientname)

We also add support for controlling the virtualenv
--system-site-packages options by passing -s to the
install-openstack-client bin. This is externally
controllable by setting the {CLIENTNAME}_EXTRA_INSTALL_OPTS
variable. (defaults to "")

This is cleaner in that we aren't hard coding /opt/stack/venv
everywhere and also supports both isolated and shared
venvs (which some people would like to use).

Change-Id: Ieb90be14a5fb1a644d253777d6468870995372b0