Add fedora-rdo-icehouse element
The fedora-rdo-icehouse element is for building Fedora images using rdo
packages from the Fedora Icehouse repository. Initially, it's just
defining environment variables to setup package install types and
requiring the fedora-rdo-icehouse-repository element to enable the yum
repository.

This element also requires the baremetal element as that element
contains a fix (stable interface names) that is needed in all Fedora
environments.

This change renames the existing fedora-rdo-icehouse element to
fedora-rdo-icehouse-repository which just enables the repository.

Change-Id: I7dbad7345ac497ed994e90fcef8076538cef1f90