Treat failure to hear back from neutron as non fatal
A recent change to nova requires neutron to have admin credentials in
order to report back to nova that VIF plugging is complete. Our neutron
doesn't have the appropriate credentials and adding them isn't trivial
as we don't know the tenant id when we do the heat stack-create of the
overcloud.

This commit should be considered a temporary workaround that disables the
requirement, this will unblock ci while we figure out how to handle the
problem correctly.

Related-Bug: #1293782

Change-Id: I1a31d31e3d3ab923217eaf3b2e865a28dd57ec44