Remove mostly empty directories
The os-apply-config and os-config-applier directories are empty other
than 1 hidden file in each. This appears to be the case just so that
they will exist when the rsync commands run to install the templates
into the image. It's possible the directories won't exist otherwise if
no elements provided any templates.

Instead of using empty directories and hidden files to force them to
exist in git, just check for the existence of each directory before
running the rsync commands.

Change-Id: I0f286a28601c292b5f70258747def36a08d93a87