Install the check_mk Agent
Check_MK is an extension to the Nagios monitoring system that
allows creating rule-based configuration using Python and
offloading work from the Nagios core to make it scale better,
allowing more systems to be monitored from a single Nagios server.

This patch provides an element to install the agent along with the
logwatch plugin to alert on logging events.  Port 6556 is also
opened in the firewall to allow an icinga server running check_mk
to connect.

More information on check_mk can be found at
http://mathias-kettner.com/check_mk.html

Change-Id: Ibc93bbb1f99efdde2792ece63f5cf332a2edddb7