Ensure the nova element is installable
On a system that has access only to a pypi-mirror without
cached wheels, there a failure to build the overcloud-compute
image.

This is due to the requisite *-dev packages not necessarily
being present on the system.

(With binary wheels, this problem won't surface.)

Correct the nova element to specify its binary dependencies
explicitly.

Change-Id: Iae97ccb63a9fd33d2aa132c0b6c245554cbadf63