Expose dnsmasq options
Expose the ability to set dnsmasq options for neutron dhcp agent. This
will allow us to configure mtu to be 1400 for tenant instances on the
overcloud.  This should help with poor network performance and vm's that
are just plain unreachable via ssh due to the GRE tunnel overhead.

This will use a corresponding change to tripleo-heat-templates:
If24326045987b5a484ba2f71f591092987966536

Change-Id: I7dba10406a8560b5e177b5a983485d637cca3105
Partial-Bug: #1270646