Prevent network interface logging w/ ovs agent
Openvswitch will create many network interfaces, especially when
supporting a nova compute node. We don't want any bugs in the underyling
OS tools to result in an explosion of network interface specific logs,
as we have seen that this can overwhelm upstart. This is mostly to be
defensive against future problems.

Change-Id: Id8310c5f24ea596ed2443bcb927a07b005980981